% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_and_sort_pValues.R
\name{adjust_and_sort}
\alias{adjust_and_sort}
\title{Adjust and Sort Pathway \eqn{p}-Values}
\usage{
adjust_and_sort(pVals_vec, genesets_ls, adjust = TRUE,
  proc_vec = c("Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH",
  "BY", "ABH", "TSBH"), ...)
}
\arguments{
\item{pVals_vec}{A named vector of permutation \eqn{p}-values returned by the
\code{\link{permTest_OmicsSurv}}, \code{\link{permTest_OmicsReg}}, or
\code{\link{permTest_OmicsCateg}} functions when the analysis performed was
AES-PCA. Otherwise, when the analysis was performed with Supervised PCA, a
named vector of \eqn{p}-values from the \code{\link{weibullMix_pValues}}
function.}

\item{genesets_ls}{A list of known gene pathways. This pathway list must
contain
\itemize{
  \item{pathways : }{A named list of character vectors where each vector
    contains the names of the genes in that specific pathway.}
  \item{TERMS : }{A character vector the same length as \code{pathways}
    containing the full pathway descriptions.}
  \item{setsize : }{An integer vector the same length as \code{pathways}
    containing the number of genes present in the pathway.}
}}

\item{adjust}{Should you adjust the \eqn{p}-values for multiple comparisons?
Defaults to TRUE.}

\item{proc_vec}{Character vector of procedures. The returned data frame will
be sorted in ascending order by the first procedure in this vector, with
ties broken by the unadjusted \eqn{p}-value. If only one procedure is
selected, then it is necessarily the first procedure.}

\item{...}{Additional arguments to pass to the \code{\link{adjustRaw_pVals}}
function.}
}
\value{
A data frame with columns
}
\description{
Adjust the pathway \eqn{p}-values, then return a data frame of
  the relevant pathway information, sorted by adjusted significance.
}
\details{
This is a wrapper function for the \code{\link{adjustRaw_pVals}}
  function.
}
\examples{
  NULL
}
