% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multtest_pvalues.R
\name{adjustRaw_pVals}
\alias{adjustRaw_pVals}
\title{Adjust p-Values for Simple Multiple-Testing Procedures}
\usage{
adjustRaw_pVals(rawp, proc = c("Bonferroni", "Holm", "Hochberg", "SidakSS",
  "SidakSD", "BH", "BY", "ABH", "TSBH"), alpha = 0.05, na.rm = FALSE,
  as.multtest.out = FALSE)
}
\arguments{
\item{rawp}{A vector of raw (unadjusted) p-values for each hypothesis under
consideration. These could be nominal p-values, for example, from t-tables,
or permutation p-values.}

\item{proc}{A vector of character strings containing the names of the
multiple testing procedures for which adjusted p-values are to be
computed. This vector should include any of the options listed in the
"Details" Section. Adjusted p-values are computed for simple FWER- and
FDR- controlling procedures based on a vector of raw (unadjusted) p-values.}

\item{alpha}{A nominal type I error rate, or a vector of error rates, used
for estimating the number of true null hypotheses in the two-stage
Benjamini & Hochberg procedure (\code{"TSBH"}). Default is 0.05.}

\item{na.rm}{An option for handling NA values in a list of raw p-values. If
\code{FALSE}, the number of hypotheses considered is the length of the
vector of raw p-values. Otherwise, if \code{TRUE}, the number of
hypotheses is the number of raw p-values which were not \code{NA}s.}

\item{as.multtest.out}{Should the output match the output from the
\code{mt.rawp2adjp} function? If not, the output will match the input (a
vector). Defaults to \code{FALSE}.}
}
\value{
A vector of the same length and order as \code{rawp}, unless the user
   specifies that the output should match the output from the \code{multtest}
   package. In that case, the use should specify \code{as.multtest.out = TRUE}
   and this function will return output identical to that of the
   \code{mt.rawp2adjp} function from package \code{multtest}. That output is
   as follows:
   \itemize{
     \item{\code{adjp} : }{A matrix of adjusted p-values, with rows
        corresponding to hypotheses and columns to multiple testing
        procedures. Hypotheses are sorted in increasing order of their raw
        (unadjusted) p-values.}
     \item{\code{index} : }{A vector of row indices, between 1 and
        \code{length(rawp)}, where rows are sorted according to their raw
        (unadjusted) p-values. To obtain the adjusted p-values in the
        original data order, use \code{adjp\[order(index),\]}.}
     \item{\code{h0.ABH} : }{The estimate of the number of true null
        hypotheses as proposed by Benjamini & Hochberg (2000) used when
        computing adjusted p-values for the \code{"ABH"} procedure (see
        Dudoit et al., 2007).}
     \item{\code{h0.TSBH} : }{The estimate (or vector of estimates) of the
        number of true null hypotheses as proposed by Benjamini et al. (2006)
        when computing adjusted p-values for the \code{"TSBH"} procedure
        (see Dudoit et al., 2007).}
   }
}
\description{
This is the \code{mt.rawp2adjp} function from the Bioconductor
   package \code{multtest}. We did not write this function. For more
   information, see
   \url{https://www.bioconductor.org/packages/3.7/bioc/manuals/multtest/man/multtest.pdf}
}
\details{
This function computes adjusted p-values for simple multiple testing
   procedures from a vector of raw (unadjusted) p-values. The procedures
   include the Bonferroni, Holm (1979), Hochberg (1988), and Sidak procedures
   for strong control of the family-wise Type I error rate (FWER), and the
   Benjamini & Hochberg (1995) and Benjamini & Yekutieli (2001) procedures
   for (strong) control of the false discovery rate (FDR). The less
   conservative adaptive Benjamini & Hochberg (2000) and two-stage Benjamini
   & Hochberg (2006) FDR-controlling procedures are also included.

   The \code{proc} options are
   \itemize{
     \item{\code{"Bonferroni"} : }{Bonferroni single-step adjusted p-values
        for strong control of the FWER.}
     \item{\code{"Holm"} : }{Holm (1979) step-down adjusted p-values for
        strong control of the FWER.}
     \item{\code{"Hochberg"} : }{Hochberg (1988) step-up adjusted p-values
        for strong control of the FWER (for raw (unadjusted) p-values
        satisfying the Simes inequality).}
     \item{\code{"SidakSS"} : }{Sidak single-step adjusted p-values for
        strong control of the FWER (for positive orthant dependent test
        statistics).}
     \item{\code{"SidakSD"} : }{Sidak step-down adjusted p-values for strong
        control of the FWER (for positive orthant dependent test statistics).}
     \item{\code{"BH"} : }{Adjusted p-values for the Benjamini & Hochberg
        (1995) step-up FDR-controlling procedure (independent and positive
        regression dependent test statistics).}
     \item{\code{"BY"} : }{Adjusted p-values for the Benjamini & Yekutieli
        (2001) step-up FDR-controlling procedure (general dependency
        structures).}
     \item{\code{"ABH"} : }{Adjusted p-values for the adaptive Benjamini &
        Hochberg (2000) step-up FDR-controlling procedure. This method
        ammends the original step-up procedure using an estimate of the
        number of true null hypotheses obtained from p-values.}
     \item{\code{"TSBH"} : }{Adjusted p-values for the two-stage Benjamini &
        Hochberg (2006) step-up FDR-controlling procedure. This method
        ammends the original step-up procedure using an estimate of the
        number of true null hypotheses obtained from a first-pass application
        of \code{"BH"}. The adjusted p-values are a-dependent, therefore
        alpha must be set in the function arguments when using this
        procedure.}
   }
}
\examples{
   NULL
}
\author{
Sandrine Dudoit, \url{http://www.stat.berkeley.edu/~sandrine}

Yongchao Ge, \url{yongchao.ge@mssm.edu}

Houston Gilbert, \url{http://www.stat.berkeley.edu/~houston}
}
