% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AESPCA.R
\name{aes.pca}
\alias{aes.pca}
\title{Adaptive, Elastic-Net PCA}
\usage{
aes.pca(x, n, d = 1, type = c("predictor", "Gram"), max.iter = 10,
  method = "SPCA", adaptive = TRUE, para = NULL, corr = FALSE,
  seed = NULL, lambda = 1e-04)
}
\arguments{
\item{lambda}{}
}
\value{
What does the function return?
}
\description{
A function to perform adaptive, elastic-net principal component
  analysis (AES-PCA).
}
\details{
A thorough explination of how the function works
}
\examples{
   NULL
}
