% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOmics_CheckPathwayCollection.R
\name{CheckPwyColl}
\alias{CheckPwyColl}
\title{Check an Input Pathway Collection}
\usage{
CheckPwyColl(pwyColl_ls)
}
\arguments{
\item{pwyColl_ls}{A pathway collection supplied to the
\code{\link{CreateOmics}} function}
}
\value{
The same pathway collection, but with names modified as described in
   "Details" and the number of genes per pathway as the \code{setsize}
   element of the collection object.
}
\description{
Check the class and names of a \code{pathwayCollection} object.
   Add or fix names as appropriate. Add the \code{setsize} vector to the
   object.
}
\details{
If there are no names, create them. If there are missing names,
   label them. If there are duplicated names (because R is stupid and allows
   duplicate element names in a list---but not a data frame!), then use the
   \code{data.frame} name rule to append a period followed by integers to
   the end of the name string.
}
\examples{
 # DO NOT CALL THIS FUNCTION DIRECTLY. CALL FROM WITHIN CreateOmics().
}
\keyword{internal}
