% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_colonSubset.R
\docType{data}
\name{colonSurv_df}
\alias{colonSurv_df}
\title{Colon Cancer -Omics Data}
\format{A subset of a data frame with 656 genes measured on 250 subjects.
  The first two columns are the Overall Survival time (\code{OS_time}) and
  death indicator (\code{OS_event}).}
\source{
Xi Steven Chen
}
\usage{
colonSurv_df
}
\description{
Subset of colon cancer surival data set
}
\keyword{datasets}
