% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_model_CoxPH.R
\name{coxTrain_fun}
\alias{coxTrain_fun}
\title{Train Cox Proportional Hazards Model}
\usage{
coxTrain_fun(x, y, censoring.status, s0.perc)
}
\arguments{
\item{x}{A "tall" pathway data frame ($p * n$)}

\item{y}{A response vector of follow-up / event times}

\item{censoring.status}{A censoring vector}

\item{s0.perc}{A stabilization parameter. This is an argument to each of the
functions called internally.}
}
\value{
A list containing:
\itemize{
  \item{tt : }{The scaled p-dimensional score vector: each value has been
    divided by the respective standard deviation plus the \code{fudge} value.}
  \item{numer : }{The original p-dimensional score vector. From the internal
    \code{.coxscor} function.}
  \item{sd : }{The standard deviations of the scores. From the internal
    \code{.coxvar} function.}
  \item{fudge : }{A regularization scalar added to the standard deviation.
    If \code{s0.perc} is supplied, \code{fudge = quantile(sd, s0.perc)}.}
}
}
\description{
Main and utility functions for training the Cox PH model
}
\details{
See \url{https://web.stanford.edu/~hastie/Papers/spca_JASA.pdf},
   Section 5, for a description of Supervised PCA applied to survival data.
   The internal functions defined in this file are not called anywhere else,
   other than in the \code{coxTrain_fun} function itself.
}
\examples{
   NULL
}
