% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Omics_Wrapper.R
\name{create_Omics}
\alias{create_Omics}
\title{Generation Wrapper function for \code{-Omics*}-class objects}
\usage{
create_Omics(assayData_df, pathwaySet_ls, response = NULL,
  respType = c("none", "survival", "regression", "categorical"))
}
\arguments{
\item{assayData_df}{An \eqn{N \times p} data frame with named columns.}

\item{pathwaySet_ls}{A \code{pathwaySet} list of known gene pathways with two
elements:
\itemize{
  \item{\code{pathways} : }{A named list of character vectors. Each vector
     contains the names of the individual genes within that pathway as a
     vector of character strings. The names contained in these vectors must
     have non-empty overlap with the \emph{column names} of the
     \code{assayData_df} data frame. The names of the pathways (the list
     elements themselves) should be the a shorthand representation of the
     full pathway name.}
  \item{\code{TERMS}: }{ A character vector the same length as the
     \code{pathways} list with the proper names of the pathways.}
}}

\item{response}{An optional response object. See "Details" for more
information. Defaults to \code{NULL}.}

\item{respType}{What type of response has been supplied. Options are
\code{"none"}, \code{"survival"}, \code{"regression"}, and
\code{"categorical"}. Defaults to \code{"none"} to match the default
\code{response = NULL} value.}
}
\value{
A valid object of class \code{OmicsPathway}, \code{OmicsSurv},
  \code{OmicsReg}, or \code{OmicsCateg}.
}
\description{
This function calls the \code{\link{create_OmicsPath}},
   \code{\link{create_OmicsSurv}}, \code{\link{create_OmicsReg}}, and
   \code{\link{create_OmicsCateg}} functions to create valid objects of the
   classes \code{OmicsPathway}, \code{OmicsSurv}, \code{OmicsReg}, or
   \code{OmicsCateg}, respectively.
}
\details{
This function is a wrapper around the four \code{create_Omics*}
   functions, and as such is not necessary for analysis. This function simply
   exists to make \code{Omics*}-object creation easier. The values supplied
   to the \code{response} function argument can be in a list, data frame,
   matrix, vector, \code{\link[survival]{Surv}} object, or any class which
   extends these. Because this function makes "best guess" type conversions
   based on the \code{respType} argument, this argument is mandatory if
   \code{response} is non-\code{NULL}. Further, it is the responsibility of
   the user to ensure that the coerced response contained in the resulting
   code{Omics} object accurately reflects the supplied response.
}
\examples{
\dontrun{
  ###  Load the Example Data  ###
  data("colonSurv_df")
  data("colon_pathwaySet")

  ###  Create an OmicsPathway Object  ###
  colon_OmicsPath <- create_Omics(assayData_df = colonSurv_df[, -(1:2)],
                                  pathwaySet_ls = colon_pathwaySet)

  ###  Create an OmicsSurv Object  ###
  colon_OmicsSurv <- create_Omics(assayData_df = colonSurv_df[, -(1:2)],
                                  pathwaySet_ls = colon_pathwaySet,
                                  response = colonSurv_df[, 1:2],
                                  respType = "surv")

  ###  Create an OmicsReg Object  ###
  colon_OmicsReg <- create_Omics(assayData_df = colonSurv_df[, -(1:2)],
                                 pathwaySet_ls = colon_pathwaySet,
                                 response = colonSurv_df$OS_time,
                                 respType = "reg")

  ###  Create an OmicsCateg Object  ###
  colon_OmicsCateg <- create_Omics(assayData_df = colonSurv_df[, -(1:2)],
                                   pathwaySet_ls = colon_pathwaySet,
                                   response = colonSurv_df$OS_event,
                                   respType = "cat")
}

}
\seealso{
\code{\link[=OmicsPathway-class]{OmicsPathway}},
   \code{\link{create_OmicsPath}},
   \code{\link[=OmicsSurv-class]{OmicsSurv}},
   \code{\link{create_OmicsSurv}},
   \code{\link[=OmicsCateg-class]{OmicsCateg}},
   \code{\link{create_OmicsCateg}}
   \code{\link[=OmicsReg-class]{OmicsReg}}, and
   \code{\link{create_OmicsReg}}.
}
