% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Omics_All.R
\name{create_OmicsPath}
\alias{create_OmicsPath}
\alias{create_OmicsSurv}
\alias{create_OmicsReg}
\alias{create_OmicsCateg}
\title{Generation Functions for -Omics.* Class Objects}
\usage{
create_OmicsPath(massSpec_df, pathwaySet_ls)

create_OmicsSurv(massSpec_df, pathwaySet_ls, eventTime_vec, eventObserved_vec)

create_OmicsReg(massSpec_df, pathwaySet_ls, response_num)

create_OmicsCateg(massSpec_df, pathwaySet_ls, response_fact)
}
\arguments{
\item{massSpec_df}{An $N x p$ data frame with named columns}

\item{pathwaySet_ls}{A list of known gene pathways}

\item{eventTime_vec}{A numeric vector with $N$ observations corresponding to
the last observed time of follow up}

\item{eventObserved_vec}{A logical vector with $N$ observations indicating
right censoring. The values will be FALSE if the observation was censored
(i.e., we did not observe an event).}

\item{response_num}{A numeric vector of length $N$: the dependent variable
in a regression exercise}

\item{response_fact}{A factor vector of length $N$: the dependent variable of a
generalized linear model}
}
\value{
An object of classes "OmicsPathway", "OmicsSurv", "OmicsReg", or
  "OmicsCateg".
}
\description{
These functions create valid objects of the "OmicsSurv", "OmicsReg", and
  "OmicsCateg" classes.
}
\section{OmicsPathway}{

Valid OmicsPathawy objects will have no response information, just the mass
  spectroscopy (gene "design") matrix and the pathway list. OmicsPathway
  objects should be created only when unsupervised pathway extraction is
  needed. Because of the missing response, no pathway testing can be
  performed on an OmicsPathway object.
}

\section{OmicsSurv}{

Valid OmicsSurv objects will have two response vectors: a vector of the most
  recently recorded follow-up times and a logical vector if that time marks
  an event (TRUE = observed event; FALSE = right-censored observation).
}

\section{OmicsReg and OmicsCateg}{

Valid OmicsReg and OmicsCateg objects with have one response vector of
  continuous or categorial (as a factor) observations, respectively.
}

