% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesPC_extract_OmicsPath_PCs.R
\docType{methods}
\name{extract_aesPCs}
\alias{extract_aesPCs}
\alias{extract_aesPCs,OmicsPathway-method}
\title{Extract AES-PCs from expressed pathway-subsets of a mass spectrometry or
  bio-assay data frame}
\usage{
extract_aesPCs(object, numPCs = 1, parallel = FALSE, numCores = NULL, ...)

\S4method{extract_aesPCs}{OmicsPathway}(object, numPCs = 1,
  parallel = FALSE, numCores = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{OmicsPathway}.}

\item{numPCs}{The number of PCs to extract from each pathway. Defaults to 1.}

\item{parallel}{Should the computation be completed in parallel? Defaults to
\code{FALSE}.}

\item{numCores}{If \code{parallel = TRUE}, how many cores should be used for
computation? Defaults to \code{NULL}.}

\item{...}{Dots for additional internal arguments (currently unused).}
}
\value{
A list of matrices. Each element of the list will be named
  by its pathway, and the elements will be \eqn{N \times} \code{numPCs}
  matrices containing the first \code{numPCs} principal components from each
  pathway. See "Details" for more information.
}
\description{
Given a clean \code{OmicsPath} object (cleaned by the
  \code{\link{expressedOmes}} function), extract the first principal
  components from each expressed pathway in the assay design matrix.
}
\details{
This function takes in a data frame with named columns and a pathway
  list as an \code{OmicsPathway} object which has had unexpressed -Omes
  removed by the \code{\link{expressedOmes}} function. This function will
  then iterate over the list of pathways, extracting columns from the assay
  design matrix which match the genes listed in that pathway as a sub-matrix
  (as a \code{data.frame} object). This function will then call the
  \code{\link{aespca}} on each data frame in the list of pathway-specific
  design matrices, extracting the first \code{numPCs} AES principal
  components from each pathway data frame. These PC matrices are returned as
  a named list.

  NOTE: some genes will be included in more than one pathway, so these
  pathways are not mutually exclusive. Further note that there may be many
  genes in the assay design matrix that are not included in the pathways,
  so these will not be extracted to the list. It is then vitally important to
  use either a very broad and generic list of pathways or a pathways list
  that is compatible to the assay data supplied.
}
\examples{
  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use AESPCA_pVals() instead

}
\seealso{
\code{\link{create_OmicsPath}}; \code{\link{expressedOmes}};
  \code{\link{aespca}}
}
