% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_genesets.R
\docType{data}
\name{genesets_ls}
\alias{genesets_ls}
\title{Gene Pathway Sets}
\format{A list of three elements:
\itemize{
  \item{pathways : }{A list of 1329 character vectors. Each vector contains
    the names of the individual genes within that pathway as a vector of
    character strings.}
  \item{TERMS : }{A character vector of length 1329 containing the names of
    the gene pathways.}
  \item{setsize : }{A named integer vector of length 1329 containing the
    number of genes in each gene pathway.}
}}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
genesets_ls
}
\description{
A Canonical Pathways Gene Sets from the Broad Institute: File:
  c2.cp.v6.0.symbols.gmt.
}
\keyword{datasets}
