% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrixRoot.R
\name{matrixRoot}
\alias{matrixRoot}
\title{Positive root of a symmetric matrix}
\usage{
matrixRoot(x, root = 2)
}
\arguments{
\item{x}{A symmetric (necessarily square) matrix.}

\item{root}{A positive real number.}
}
\value{
A matrix, that when multiplied by itself \code{root} times, yields
  \code{x}.
}
\description{
Calculate the matrix root of a symmetric matrix via the Spectral
  Decomposition.
}
\details{
This function decomposes \code{x} into \eqn{V \times D \times V^T}
  via the \code{\link[base]{eigen}} function, sets any numerically negative
  eigenvalues to 0, calculates the root of these eigenvalues as \eqn{D^r},
  then returns the matrix \eqn{V \times D^r \times V^T}.

  See \url{https://en.wikipedia.org/wiki/Eigendecomposition_of_a_matrix}.
}
\examples{
  X <- matrix(rnorm(25), ncol = 5);   xTx <- t(X) \%*\% X
  matrixRoot(xTx)
  matrixRoot(xTx, root = 3)
}
