% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_model_OLS.R
\name{olsTrain_fun}
\alias{olsTrain_fun}
\title{Gene-Specific Regularized Ordinary Least Squares Estimates}
\usage{
olsTrain_fun(x, y, s0.perc = NULL)
}
\arguments{
\item{x}{An \eqn{p \times n} predictor matrix}

\item{y}{A response vector}

\item{s0.perc}{Percentile of the standard error of the slope estimate to be
used for regularization. The Default value of \code{NULL} will use the
median of this distribution.}
}
\value{
A list of OLS model statistics:
\itemize{
  \item{tt : }{The Student's \eqn{t} test statistic the slopes (\eqn{\beta}).}
  \item{numer : }{The estimate of \eqn{\beta}.}
  \item{sd : }{The standard error of the estimates for \eqn{\beta} (the
  standard error divided by the square root of Sxx).}
  \item{fudge : }{A regularization parameter. See Details for description.}
}
}
\description{
Model statistics for Ordinary Least Squares (OLS) regression by
  gene
}
\details{
This function calculates the Sxx, Syy, and Sxy sums from the gene-
  specific OLS models, then calculates estimates of the regression slopes for
  each gene and their corresponding regularized test statistics,
  \deqn{t = \hat{\beta} / (sd + e),}
  where \eqn{e} is a regularization parameter.

  If \code{s0.perc} is \code{NULL}, then \eqn{e} is median of the \code{sd}
  values. Otherwise, \eqn{e} is set equal to \code{quantile(sd, s0.perc)}.
}
\examples{
  NULL
}
