% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClass_OmicsCateg.R
\docType{class}
\name{OmicsCateg-class}
\alias{OmicsCateg-class}
\alias{create_OmicsCateg}
\title{An S4 class for categorical responses within an "OmicsPathway" object}
\description{
This creates the "OmicsCateg" class which extends the
  "OmicsPathway" master class.
}
\section{Slots}{

\describe{
\item{\code{massSpec}}{An $N x p$ data frame with named columns}

\item{\code{pathwaySet}}{A list of known gene pathways}

\item{\code{response}}{A factor vector of length $N$: the dependent variable of a
generalized linear model}
}}

\seealso{
\code{"\link[=OmicsPathway-class]{OmicsPathway}"}
}
