% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClass_OmicsPath.R
\docType{class}
\name{OmicsPathway-class}
\alias{OmicsPathway-class}
\alias{create_OmicsPath}
\title{An S4 class for mass spectronomy data and gene pathway lists}
\description{
An S4 class for mass spectronomy data and gene pathway lists
}
\section{Slots}{

\describe{
\item{\code{massSpec}}{An $N x p$ data frame with named columns}

\item{\code{pathwaySet}}{A list of known gene pathways}
}}

