% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClass_OmicsReg.R
\docType{class}
\name{OmicsReg-class}
\alias{OmicsReg-class}
\title{An S4 class for continuous responses within an "OmicsPathway" object}
\description{
This creates the "OmicsReg" class which extends the
  "OmicsPathway" master class.
}
\section{Slots}{

\describe{
\item{\code{massSpec}}{An $N x p$ data frame with named columns}

\item{\code{pathwaySet}}{A list of known gene pathways}

\item{\code{response}}{A numeric vector of length $N$: the dependent variable in a
regression exercise}
}}

\seealso{
\code{"\link[=OmicsPathway-class]{OmicsPathway}"}
}
