% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClass_OmicsSurv.R
\docType{class}
\name{OmicsSurv-class}
\alias{OmicsSurv-class}
\title{An S4 class for survival responses within an "OmicsPathway" object}
\description{
This creates the "OmicsSurv" class which extends the
  "OmicsPathway" master class.
}
\section{Slots}{

\describe{
\item{\code{massSpec}}{An $N x p$ data frame with named columns}

\item{\code{pathwaySet}}{A list of known gene pathways}

\item{\code{eventTime}}{A numeric vector with $N$ observations corresponding to the
last observed time of follow up}

\item{\code{eventObserved}}{A logical vector with $N$ observations indicating right
censoring. The values will be FALSE if the observation was censored (i.e.,
we did not observe an event).}
}}

\seealso{
\code{"\link[=OmicsPathway-class]{OmicsPathway}"}
}
