% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_pathway_pValues.R
\name{pathway_pValues}
\alias{pathway_pValues}
\title{Calculate the \eqn{p}-values from a mixture of Weibull Extreme Value
   Distributions for supervised PCA}
\usage{
pathway_pValues(optimParams_vec, max_tScores_vec, genelist_ls,
  FDRadjust = TRUE, multTestProc = "BH")
}
\arguments{
\item{optimParams_vec}{A named vector of the estimated values for the
parameters which minimize the likelihood as returned by the function
\code{\link{weibullMix_optimParams}}.}

\item{max_tScores_vec}{A vector of the maximum absolute \eqn{t}-scores for
each pathway when under the alternative model (the response vector as is).}

\item{genelist_ls}{A list of three elements:
\itemize{
  \item{\code{pathways} : }{A list of character vectors such that each
     vector contains the ID numbers (as a character) of the individual
     genes within that pathway as a vector of character strings.}
  \item{\code{TERMS} : }{A character vector containing the names of the
     gene pathways.}
  \item{\code{setsize} : }{A named integer vector containing the number of
     genes in each gene pathway.}
}}

\item{FDRadjust}{Should the \eqn{p}-values be adjusted for multiple
comparisons? Defaults to \code{TRUE}.}

\item{multTestProc}{If the \eqn{p}-values should be adjusted, which procedure
should be used? Options are passed to the \code{\link{adjustRaw_pVals}}
function. Specify multiple procedures via \code{c(...)}. Defaults to
\code{"BH"}.}
}
\value{
A data frame with columns for the pathway names, pathway set sizes,
   raw pathway \eqn{p}-values, and a column of FDR-adjusted \eqn{p}-values
   for each adjustment method specified.
}
\description{
Calculate pathway-specific \eqn{p}-values for supervised PCA
   and their associated False Discovery Rates (FDR).
}
\details{
This function takes in the optimal parameters returned by the
   \code{\link{weibullMix_optimParams}} function, the maximum \eqn{t}-scores
   for each gene pathway, and the list of gene pathway information. This
   function will calculate the \eqn{p}-value for each \eqn{t}-score given the
   Gumbel Extreme Value mixture distribution parametrized by the values
   returned by the \code{\link{weibullMix_optimParams}} function. If
   requested, this function will also calculate the FDR associated with all
   pathway \eqn{p}-values via requested FDR-adjustment procedure. The default
   procedure is the Benjamini & Hochberg (1995) step-up FDR-controlling
   procedure, but any procedure implemented in the
   \code{\link{adjustRaw_pVals}} function is available.
}
\examples{
  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use superPCA_pVals() instead
}
