% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayPCA.R
\docType{package}
\name{pathwayPCA}
\alias{pathwayPCA}
\alias{pathwayPCA-package}
\title{Extract and Test the Significance of Pathway-Specific Principal
   Components}
\description{
The \code{pathwayPCA} package has three main components:
   \itemize{
      \item{Import and Tidy Data: }{
        \url{https://gabrielodom.github.io/pathwayPCA/articles/C2-Importing_Data.html}
      }
      \item{Create \code{Omics} Data Objects }{
        \url{https://gabrielodom.github.io/pathwayPCA/articles/C3-Create_Omics_Objects.html}
      }
      \item{Test Pathway Significance }{
        \url{https://gabrielodom.github.io/pathwayPCA/articles/C4-Methods_Walkthrough.html}
      }
    }
  For an overview of these four pieces in context, please see our Quickstart
  Guide:
  \url{https://gabrielodom.github.io/pathwayPCA/articles/C1-Quickstart_Guide.html}
}
