% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permtest_aesPCs_CoxPH.R
\docType{methods}
\name{permTest_OmicsSurv}
\alias{permTest_OmicsSurv}
\alias{permTest_OmicsSurv,OmicsSurv-method}
\title{Permutation Test for Pathway PCs for Survival Response}
\usage{
permTest_OmicsSurv(OmicsSurv, pathwayPCs_ls, numReps = 1000,
  parallel = FALSE, numCores = NULL, ...)

\S4method{permTest_OmicsSurv}{OmicsSurv}(OmicsSurv, pathwayPCs_ls,
  numReps = 1000, parallel = FALSE, numCores = NULL, ...)
}
\arguments{
\item{OmicsSurv}{A data object of class \code{OmicsSurv}, created by the
\code{\link{create_OmicsSurv}} function.}

\item{pathwayPCs_ls}{A list of pathway PC matrices returned by the
\code{\link{extract_aesPCs}} function.}

\item{numReps}{How many permuted models to fit? Defaults to 1000}

\item{parallel}{Should the comuptation be completed in parallel? Defaults to
\code{FALSE}.}

\item{numCores}{If \code{parallel = TRUE}, how many cores should be used for
computation?}

\item{...}{Dots for additional internal arguments (currently unused)}
}
\value{
A named vector of pathway permutation \eqn{p}-values.
}
\description{
Given an \code{OmicsSurv} object and a list of pathway PCs from
  the \code{\link{extract_aesPCs}} function, test if each expressed pathway
  in the MS design matrix is significantly related to the survival output.
}
\details{
This function takes in a list of the first principal components
  from each pathway and an object of class `OmicsSurv`. This function will
  then calculate the AIC of a Cox Proportional Hazards model (via the
  \code{\link[survival]{coxph}} function) with the original observations as
  response and the pathway principal components as the predictor matrix.

  Then, this function will create \code{numReps} permutations of the survival
  response, fit models to each of these premuted responses (holding the path
  predictor matrix fixed), and calculate the AIC of each model. This function
  will return a named vector of permutation \eqn{p}-values, where the value
  for each pathway is the proportion of models for which the AIC of the
  permuted response model is less than the AIC of the original model.
}
\seealso{
\code{\link{create_OmicsSurv}}; \code{\link{extract_aesPCs}};
  \code{\link[survival]{coxph}}; \code{\link{sample_Survivalresp}}
}
