% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesPC_permtest_CoxPH.R
\docType{methods}
\name{PermTestSurv}
\alias{PermTestSurv}
\alias{PermTestSurv,OmicsSurv-method}
\title{AES-PCA permutation test of survival response for pathway PCs}
\usage{
PermTestSurv(OmicsSurv, pathwayPCs_ls, numReps = 1000,
  parallel = FALSE, numCores = NULL, ...)

\S4method{PermTestSurv}{OmicsSurv}(OmicsSurv, pathwayPCs_ls,
  numReps = 1000, parallel = FALSE, numCores = NULL, ...)
}
\arguments{
\item{OmicsSurv}{A data object of class \code{OmicsSurv}, created by the
\code{\link{CreateOmicsSurv}} function.}

\item{pathwayPCs_ls}{A list of pathway PC matrices returned by the
\code{\link{ExtractAESPCs}} function.}

\item{numReps}{How many permuted models to fit? Defaults to 1000.}

\item{parallel}{Should the computation be completed in parallel? Defaults to
\code{FALSE}.}

\item{numCores}{If \code{parallel = TRUE}, how many cores should be used for
computation? Internally defaults to the number of available cores minus 2.}

\item{...}{Dots for additional internal arguments (currently unused).}
}
\value{
A named vector of pathway permutation \eqn{p}-values.
}
\description{
Given an \code{OmicsSurv} object and a list of pathway principal
   components (PCs) from the \code{\link{ExtractAESPCs}} function, test if
   each pathway with features recorded in the bio-assay design matrix is
   significantly related to the survival output.
}
\details{
This function takes in a list of the first principal components
  from each pathway and an object of class \code{OmicsSurv}. This function
  will then calculate the AIC of a Cox Proportional Hazards model (via the
  \code{\link[survival]{coxph}} function) with the original observations as
  response and the pathway principal components as the predictor matrix.

  Then, this function will create \code{numReps} permutations of the survival
  response, fit models to each of these permuted responses (holding the path
  predictor matrix fixed), and calculate the AIC of each model. This function
  will return a named vector of permutation \eqn{p}-values, where the value
  for each pathway is the proportion of models for which the AIC of the
  permuted response model is less than the AIC of the original model.
}
\examples{
  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use AESPCA_pVals() instead

}
\seealso{
\code{\link{CreateOmicsSurv}}; \code{\link{ExtractAESPCs}};
  \code{\link[survival]{coxph}}; \code{\link{SampleSurv}}
}
\keyword{internal}
