% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printClass_pathwayCollection.R
\name{print.pathwayCollection}
\alias{print.pathwayCollection}
\title{Display the Summary of a \code{pathwayCollection}-class Object.}
\usage{
\method{print}{pathwayCollection}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pathwayCollection}.}

\item{...}{Lazy dots for additional internal arguments (currently unused).}
}
\value{
\code{x}, returned invisibly (with the \code{\link{invisible}}
   function).
}
\description{
The display method for pathways lists as returned by the
   \code{\link{read_gmt}} function.
}
\details{
This function sets a \code{print} method for
   \code{pathwayCollection} objects.
}
\examples{
  ###  Load the Example Data  ###
  data("colon_pathwayCollection")

  ###  Print / Show  ###
  colon_pathwayCollection

}
\seealso{
\code{\link{read_gmt}}; \code{\link{write_gmt}}
}
\keyword{internal}
