% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printClass_pathwaySet.R
\name{print.pathwaySet}
\alias{print.pathwaySet}
\title{Display the Summary of a\code{pathwaySet}-class Object.}
\usage{
\method{print}{pathwaySet}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pathwaySet}.}

\item{...}{Lazy dots for additional internal arguments (currently unused).}
}
\value{
\code{x}, returned invisibly (with the \code{\link{invisible}}
   function).
}
\description{
The display method for gene set / pathway set lists as returned
   by the \code{\link{read_gmt}} function.
}
\details{
This function sets a \code{print} method for \code{pathwaySet}
   objects.
}
\examples{
  ###  Load the Example Data  ###
  data("colon_pathwaySet")

  ###  Print / Show  ###
  colon_pathwaySet

}
\seealso{
\code{\link{read_gmt}}
}
