% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_permuteSamples.R
\name{randomControlSample}
\alias{randomControlSample}
\alias{sample_Survivalresp}
\alias{sample_Regresp}
\alias{sample_Classifresp}
\title{Parametric Bootstrap and Non-parametric Permutations of a Response}
\usage{
sample_Survivalresp(response_vec, censor_vec, parametric = FALSE)

sample_Regresp(response_vec, parametric = FALSE)

sample_Classifresp(response_vec, parametric = FALSE)
}
\arguments{
\item{response_vec}{The dependent vector to sample from. For survival
response, this is the vector of event times. For regression or n-ary
classification, this is the vector of responses.}

\item{censor_vec}{The censoring indicator vector for survival response. This
is coded as 1 for a right-censoring occurence and 0 for a recorded event.}

\item{parametric}{Should the random sample be taken using a parametric
bootstrap sample? Defaults to \code{FALSE}.}
}
\description{
Create a random parametric bootstrap sample or a permutation of
   the input response vector or matrix (for survival outcomes).
}
\details{
The distributions (for \code{parametric = TRUE}) are Weibull for
   survival times, Normal for regression, and n-ary Multinomial for
   classification. Distributional parameters are estimated with their maximum
   likelihood estimates. When \code{parametric = FALSE}, the response vector
   or survival matrix is simply permuted by row.
}
\examples{
   NULL
   NULL
   NULL
}
