% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmt.R
\name{read_gmt}
\alias{read_gmt}
\title{Read a \code{.gmt} file}
\usage{
read_gmt(file, delim = "\\t")
}
\arguments{
\item{file}{A path to a file or a connection. This file must be a \code{.gmt}
file, otherwise input will likely be nonsense. See the "Details" section
for more information.}

\item{delim}{The \code{.gmt} delimiter. As proper \code{.gmt} files are tab
delimited, this defaults to \code{"\\t"}.}
}
\value{
A \code{pathwaySet} list of gene sets. This list has three elements:
\itemize{
  \item{\code{pathways} : }{A named list of character vectors. Each vector
     contains the names of the individual genes within that pathway as a
     vector of character strings.}
  \item{\code{TERMS} : }{A character vector the same length as the
     \code{pathways} list with the proper names of the pathways.}
  \item{\code{GSEA_link} : }{A character vector the same length as the
     \code{pathways} list with hyperlinks to the MSigDB description card for
     that pathway.}
}
}
\description{
Read a gene set file in Gene Matrix Transposed (\code{.gmt})
   format, with special performance consideration for large files.
}
\details{
This function uses \code{R}'s \code{\link{readChar}} function to
   improve character input performance over \code{\link{readLines}} (and
   far improve input performance over \code{\link{scan}}).

   See the Broad Institute's "Data Formats" page for a description of the
   Gene Matrix Transposed file format:
   \url{https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29}
}
\examples{
\dontrun{
  # If you have installed the package:
  data_path <- system.file("extdata", "c2.cp.v6.0.symbols.gmt",
                            package = "pathwayPCA", mustWork = TRUE)
  geneset_ls <- read_gmt(data_path)

  # If you are using the development version from GitHub:
  geneset_ls <- read_gmt("inst/extdata/c2.cp.v6.0.symbols.gmt")
}

}
