% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_supervised_reducedGenesets.R
\docType{data}
\name{supervised_Genesets4240_ls}
\alias{supervised_Genesets4240_ls}
\title{Reduced Gene Pathway Sets for Supervised Example}
\format{A list of three elements:
\itemize{
  \item{pathways : }{A list of 4240 character vectors. Each vector contains
    the ID numbers (as a character) of the individual genes within that
    pathway as a vector of character strings.}
  \item{TERMS : }{A character vector of length 4240 containing the names of
    the gene pathways.}
  \item{setsize : }{A named integer vector of length 4240 containing the
    number of genes in each gene pathway.}
}}
\source{
UNKNOWN - GET INFO FROM STEVEN
}
\usage{
supervised_Genesets4240_ls
}
\description{
A Gene Ontology Pathways Gene Set. This is a list of genes with
   Entrez IDs and GO annotations (\code{hgu133plus2ENTREZID} and
   \code{hgu133plus2GO} from \code{Bioconductor} package
   \code{hgu133plus2.db}, respectively), with pathways with fewer than five
   genes or more than 180 genes removed (99.5\% of all pathways we encounter
   have fewer than 180 genes).
}
\keyword{datasets}
