% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_supervised_patInfo.R
\docType{data}
\name{supervised_patInfo_df}
\alias{supervised_patInfo_df}
\title{Matched Patient Info for Tumour Geneomics Data}
\format{A data frame with 15 identifiers and responses measured on 177
  subjects (from which 177 tumors were excised). Panel data corresponding
  to the 177 subjects' tumours are included in the file
  \code{\link{supervised_Tumors_df}}.}
\source{
UNKNOWN - ASK STEVEN
}
\usage{
supervised_patInfo_df
}
\description{
Tumor host subject information for UNKNOWN tumor data set
}
\keyword{datasets}
