% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_supervised_tumors.R
\docType{data}
\name{supervised_Tumors_df}
\alias{supervised_Tumors_df}
\title{Tumour Geneomics Data}
\format{A data frame with 16688 genes measured on 177 tumours. Subject data
  corresponding to the 177 tumours are included in the file
  \code{\link{supervised_patInfo_df}}. NOTE: this data is "tall": p x n.}
\source{
UNKNOWN - ASK STEVEN
}
\usage{
supervised_Tumors_df
}
\description{
UNKNOWN
}
\keyword{datasets}
