% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tControl.R
\docType{data}
\name{tControl_mat}
\alias{tControl_mat}
\title{Null Model t-Statistic Matrix}
\format{A 7,949 x 20 matrix of t-scores. Each row corresponds to the t-scores
  calculated when the patient survival information is randomly permuted and
  then fit to the 7,949 gene pathway sets. There are 20 t-scores,
  corresponding to 20 threshold values for the Cox Propotional Hazards
  regression model.}
\source{
Personal computation
}
\usage{
tControl_mat
}
\description{
A matrix of t-scores from pathway models for the
  \code{supervised_Tumors_df} and \code{supervised_patInfo_df} data set.
}
\keyword{datasets}
