% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_assay.R
\name{transpose_assay}
\alias{transpose_assay}
\title{Transpose an Assay (Data Frame)}
\usage{
transpose_assay(assay_df, firstColIsFeatureNames = TRUE)
}
\arguments{
\item{assay_df}{A data frame with numeric values to transpose}

\item{firstColIsFeatureNames}{Are the data row names in the first column of
\code{df}? Defaults to \code{TRUE}. If not, this function assumes that the
data row names are accesible by the \code{\link{rownames}} function.}
}
\value{
The transposition of \code{df}, with row and column names preserved.
}
\description{
Transpose an object of class \code{data.frame} that contains
   assay measurements while preserving row (feature) and column (sample)
   names.
}
\details{
This function is designed to transpose "tall" assay data frames
   (where genes or proteins are the rows and patient or tumour samples are
   the columns). This function also transposes the row (feature) names to
   column names and the column (sample) names to row names. Notice that all
   rows and columns (other than the feature name column, as applicable) are
   numeric.
}
\examples{
   x_mat <- matrix(rnorm(5000), ncol = 20, nrow = 250)
   rownames(x_mat) <- paste0("gene_", 1:250)
   colnames(x_mat) <- paste0("sample_", 1:20)
   x_df <- as.data.frame(x_mat, row.names = rownames(x_mat))

   transpose_assay(x_df, firstColIsFeatureNames = FALSE)

}
