% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tScores4240.R
\docType{data}
\name{tScores4240_mat}
\alias{tScores4240_mat}
\title{Alternative Model t-Statistic Matrix}
\format{A 4,240 x 20 matrix of t-scores. Each row corresponds to the t-scores
  calculated when the patient survival information is fit to the 4,240 gene
  pathway sets. There are 20 t-scores, corresponding to 20 threshold values
  for the Cox Propotional Hazards regression model.}
\source{
Personal computation
}
\usage{
tScores4240_mat
}
\description{
A matrix of t-scores from a subset of the pathway models for
  the \code{supervised_Tumors_df} and \code{supervised_patInfo_df} data set.
  These pathways are a subset of the full pathway set where pathways smaller
  than 5 genes have been deleted.
}
\keyword{datasets}
