% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesPC_wrapper.R
\docType{methods}
\name{AESPCA_pVals}
\alias{AESPCA_pVals}
\alias{AESPCA_pVals,OmicsPathway-method}
\title{Test Pathways with AES-PCA}
\usage{
AESPCA_pVals(object, numPCs = 1, min.features = 3, numReps = 1000,
  parallel = FALSE, numCores = NULL, adjustpValues = TRUE,
  adjustment = c("Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH",
  "BY", "ABH", "TSBH"), ...)

\S4method{AESPCA_pVals}{OmicsPathway}(object, numPCs = 1, min.features = 3,
  numReps = 1000, parallel = FALSE, numCores = NULL,
  adjustpValues = TRUE, adjustment = c("Bonferroni", "Holm", "Hochberg",
  "SidakSS", "SidakSD", "BH", "BY", "ABH", "TSBH"), ...)
}
\arguments{
\item{object}{An object of class \code{OmicsPathway} with a response matrix}

\item{numPCs}{The number of PCs to extract from each pathway. Defaults to 1.}

\item{min.features}{What is the smallest number of genes allowed in each
pathway? This argument must be kept constant across all calls to this
function which use the same pathway list. Defaults to 3.}

\item{numReps}{The number of permutations to take of the data to calculate a
\eqn{p}-value for each pathway. Defaults to 1000.}

\item{parallel}{Should the comuptation be completed in parallel? Defaults to
\code{FALSE}.}

\item{numCores}{If \code{parallel = TRUE}, how many cores should be used for
computation?}

\item{adjustpValues}{Should you adjust the \eqn{p}-values for multiple
comparisons? Defaults to TRUE.}

\item{adjustment}{Character vector of procedures. The returned data frame
will be sorted in ascending order by the first procedure in this vector,
with ties broken by the unadjusted \eqn{p}-value. If only one procedure is
selected, then it is necessarily the first procedure. See the documentation
for the \code{\link{adjustRaw_pVals}} function for the adjustment procedure
definitions and citations.}

\item{...}{Dots for additional internal arguments}
}
\value{
A data frame with columns
\itemize{
  \item{\code{pathways} : }{The names of the pathways in the \code{Omics*}}
    object (stored in \code{object@pathwaySet$pathways})
  \item{\code{setsize} : }{The number of genes in each of the original
    pathways (as stored in the \code{object@pathwaySet$setsize} object)}
  \item{\code{terms} : }{The pathway description, as stored in the
    \code{object@pathwaySet$TERMS} object}
  \item{\code{rawp} : }{The unadjusted \eqn{p}-values of each pathway}
  \item{\code{...} : }{Additional columns as specified through the
    \code{adjustment} argument}
}

The data frame will be sorted in ascending order by the method specified
  first in the \code{adjustment} argument. If \code{adjustpValues = FALSE},
  then the data frame will be sorted by the raw \eqn{p}-values. If you have
  the suggested \code{tidyverse::} package suite loaded, then this data frame
  will print as a \code{\link[tibble]{tibble}}. Otherwise, it will stay a
  simple data frame.
}
\description{
Given a supervised \code{OmicsPath} object (one of
  \code{OmicsSurv}, \code{OmicsReg}, or \code{OmicsCateg}), extract the first
  adaptive, elastic-net, sparse principal components from each expressed
  pathway in the MS design matrix, test their association with the response
  matrix, and return a data frame of the adjusted \eqn{p}-values for each
  pathway.
}
\details{
This is a wrapper function for the \code{\link{expressedOmes}},
  \code{\link{extract_aesPCs}}, \code{\link{permTest_OmicsSurv}},
  \code{\link{permTest_OmicsReg}}, and \code{\link{permTest_OmicsCateg}}
  functions.
}
\seealso{
\code{\link{expressedOmes}}; \code{\link{create_OmicsPath}};
  \code{\link{create_OmicsSurv}}; \code{\link{create_OmicsReg}};
  \code{\link{create_OmicsCateg}}; \code{\link{extract_aesPCs}};
  \code{\link{permTest_OmicsSurv}}; \code{\link{permTest_OmicsReg}};
  \code{\link{permTest_OmicsCateg}}; \code{\link{adjust_and_sort}}
}
