% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_colonSubset.R
\docType{data}
\name{colonSurv_df}
\alias{colonSurv_df}
\title{Colon Cancer -Omics Data}
\format{
A subset of a data frame containing 656 of 2022 genes measured on
  250 subjects. The first two columns are the Overall Survival time
  (\code{OS_time}) and death indicator (\code{OS_event}).
}
\source{
GEO GSE17538
  \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17538}
}
\usage{
data(colonSurv_df)
}
\description{
Subset of a colon cancer survival data set, with subject
  response and assay values.
}
\keyword{datasets}
