% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Omics_All.R
\name{create_OmicsPath}
\alias{create_OmicsPath}
\alias{create_OmicsSurv}
\alias{create_OmicsReg}
\alias{create_OmicsCateg}
\title{Generation functions for \code{-Omics*}-class objects}
\usage{
create_OmicsPath(assayData_df, pathwayCollection_ls)

create_OmicsSurv(assayData_df, pathwayCollection_ls, eventTime_num,
  eventObserved_lgl)

create_OmicsReg(assayData_df, pathwayCollection_ls, response_num)

create_OmicsCateg(assayData_df, pathwayCollection_ls, response_fact)
}
\arguments{
\item{assayData_df}{An \eqn{N \times p} data frame with named columns.}

\item{pathwayCollection_ls}{A \code{pathwayCollection} list of known gene
pathways with two or three elements:
\itemize{
  \item{\code{pathways} : }{A named list of character vectors. Each vector
     contains the names of the individual genes within that pathway as a
     vector of character strings. The names contained in these vectors must
     have non-empty overlap with the \emph{column names} of the
     \code{assayData_df} data frame. The names of the pathways (the list
     elements themselves) should be the a shorthand representation of the
     full pathway name.}
  \item{\code{TERMS}: }{A character vector the same length as the
     \code{pathways} list with the proper names of the pathways.}
  \item{\code{description} : }{An optional character vector the same length
     as the \code{pathways} list with additional information about the
     pathways.}
}}

\item{eventTime_num}{A \code{numeric} vector with \eqn{N} observations
corresponding to the last observed time of follow up.}

\item{eventObserved_lgl}{A \code{logical} vector with \eqn{N} observations
indicating right-censoring. The values will be \code{FALSE} if the
observation was censored (i.e., we did not observe an event).}

\item{response_num}{A \code{numeric} vector of length \eqn{N}: the dependent
variable in an ordinary regression exercise.}

\item{response_fact}{A \code{factor} vector of length \eqn{N}: the dependent
variable of a generalized linear regression exercise.}
}
\value{
A valid object of class \code{OmicsPathway}, \code{OmicsSurv},
  \code{OmicsReg}, or \code{OmicsCateg}.
}
\description{
These functions create valid objects of class \code{OmicsPathway},
  \code{OmicsSurv}, \code{OmicsReg}, or \code{OmicsCateg}.
}
\details{
Please note that the classes of the parameters are \emph{not}
  flexible. The -Omics assay data \emph{must} be or extend the class
  \code{data.frame}, and the response values (for a survival, regression,
  or classification object) \emph{must} match their expected classes
  \emph{exactly}. The reason for this is to encourage the end user to pay
  attention to the quality and format of their input data. Because the
  functions internal to this package have only been tested on the classes
  described in the Arguments section, these class checks prevent unexpected
  errors (or worse, incorrect computational results without an error). These
  draconian input class restrictions protect the accuracy of your data
  analysis.

  Also note the following: if the supplied \code{pathways} object within your
  \code{pathwayCollection_ls} list has no names, then this pathway list will
  be named \code{path1}, \code{path2}, \code{path3}, ...; if any of the
  pathways are missing names, then the missing pathways will be named
  \code{noName} followed by the index of the pathway. For example, if the
  112th pathway in the \code{pathways} list has no name (but other pathways
  do), then this pathway will be named \code{noName112}. Furthermore, if any
  of the pathway names are duplicated, then the duplicates will have
  \code{.1}, \code{.2}, \code{.3}, ... appended to the duplicate names until
  all pathway names are unique. Once all pathways have been verified to have
  unique names, then the pathway names are attached as attributes to the
  \code{TERMS} and \code{setsize} vectors (the \code{setsize} vector is
  calculated at object creation).

  If your gene pathways list is stored in a \code{.gmt} file, use the
  \code{\link{read_gmt}} function to import your pathways list as a
  \code{pathwayCollection} list object.
}
\section{OmicsPathway}{

Valid \code{OmicsPathway} objects will have no response information, just the
  mass spectrometry or bio-assay ("design") matrix and the pathway list.
  \code{OmicsPathway} objects should be created only when unsupervised
  pathway extraction is needed (not possible with Supervised PCA). Because of
  the missing response, no pathway testing can be performed on an
  \code{OmicsPathway} object.
}

\section{OmicsSurv}{

Valid \code{OmicsSurv} objects will have two response vectors: a vector of
  the most recently recorded follow-up times and a logical vector if that
  time marks an event (\code{TRUE}: observed event; \code{FALSE}: right-
  censored observation).
}

\section{OmicsReg and OmicsCateg}{

Valid \code{OmicsReg} and \code{OmicsCateg} objects with have one response
  vector of continuous (\code{numeric}) or categorial (\code{factor})
  observations, respectively.
}

\examples{
\dontrun{
  ###  Load the Example Data  ###
  data("colonSurv_df")
  data("colon_pathwayCollection")

  ###  Create an OmicsPathway Object  ###
  colon_OmicsPath <- create_OmicsPath(
    assayData_df = colonSurv_df[, -(1:2)],
    pathwayCollection_ls = colon_pathwayCollection
  )

  ###  Create an OmicsSurv Object  ###
  colon_OmicsSurv <- create_OmicsSurv(
    assayData_df = colonSurv_df[, -(1:2)],
    pathwayCollection_ls = colon_pathwayCollection,
    eventTime_num = colonSurv_df$OS_time,
    eventObserved_lgl = as.logical(colonSurv_df$OS_event)
  )

  ###  Create an OmicsReg Object  ###
  colon_OmicsReg <- create_OmicsReg(
    assayData_df = colonSurv_df[, -(1:2)],
    pathwayCollection_ls = colon_pathwayCollection,
    response_num = colonSurv_df$OS_time
  )

  ###  Create an OmicsCateg Object  ###
  colon_OmicsCateg <- create_OmicsCateg(
    assayData_df = colonSurv_df[, -(1:2)],
    pathwayCollection_ls = colon_pathwayCollection,
    response_fact = as.factor(colonSurv_df$OS_event)
  )
}

}
\seealso{
\code{\link[=OmicsPathway-class]{OmicsPathway}},
  \code{\link[=OmicsSurv-class]{OmicsSurv}},
  \code{\link[=OmicsReg-class]{OmicsReg}},  and
  \code{\link[=OmicsCateg-class]{OmicsCateg}}
}
