% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePathwayCollection.R
\name{CreatePathwayCollection}
\alias{CreatePathwayCollection}
\title{Manually Create a \code{pathwayCollection}-class Object.}
\usage{
CreatePathwayCollection(pathways, TERMS, ...)
}
\arguments{
\item{pathways}{A named list of character vectors. Each vector should contain
the names of the individual genes or proteins within that pathway as a
vector of character strings. The names contained in these vectors should
have non-empty overlap with the feature names of the assay data frame that
will be paired with this list in the subsequent analysis. The names of the
pathways (the list elements themselves) should be the a shorthand
representation of the full pathway name, but this is not required.}

\item{TERMS}{A character vector the same length as the \code{pathways} list
with the proper names of the pathways.}

\item{...}{Additional vectors or data components related to the
\code{pathways} list. These values should be passed as a name-value pair.
See "Details" for more information.}
}
\value{
A list object with class \code{pathwayCollection}.
}
\description{
Manually create a \code{pathwayCollection} list similar to the
   output of the \code{\link{read_gmt}} function.
}
\details{
This function checks the pathwy list and pathway term inputs and
   then creates a \code{pathwayCollection} object from them. Pass additional
   list elements (such as the \code{description} of each pathway) using the
   form \code{tag = value} through the \code{...} argument (as in the
   \code{\link{list}} function). Because some functions in the
   \code{pathwayPCA} package add and edit elements of \code{pathwayCollection}
   objects, please do not create \code{pathwayCollection} list items named
   \code{setsize} or \code{trim_setsize}.
}
\examples{
  data("colon_pathwayCollection")

  CreatePathwayCollection(
    pathways = colon_pathwayCollection$pathways,
    TERMS = colon_pathwayCollection$TERMS
  )

}
\seealso{
\code{\link{read_gmt}}
}
