% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetExpressed-omes.R
\docType{methods}
\name{expressedOmes}
\alias{expressedOmes}
\alias{expressedOmes,OmicsPathway-method}
\title{Extract expressed -omes matching a gene set from a mass spectrometry matrix}
\usage{
expressedOmes(object, trim = 3, message = TRUE, ...)

\S4method{expressedOmes}{OmicsPathway}(object, trim = 3, message = TRUE,
  ...)
}
\arguments{
\item{object}{An object of class \code{OmicsPathway}, \code{OmicsSurv},
\code{OmicsReg}, or \code{OmicsCateg}}

\item{trim}{The minimum cutoff of expressed -ome measures before a pathway
is excluded. Defaults to 3.}

\item{message}{Should this function return diagnostic messages? Messages
concern the percentage of genes included in the pathway set but not
measured in the data, genes measured in the data but not called for in the
pathways, and the number of pathways ignored due to too few number of genes
present after trimming. Defaults to \code{TRUE}.}

\item{...}{Dots for additional internal arguments (as necessary)}
}
\value{
A valid \code{Omics*}-class object. This output object will be
  identical to the input object, except that any genes present in the
  pathways list, but not present in the MS design matrix, will have been
  removed. Additionally, the pathway list will have the number of genes in
  each trimmed pathway stored as the \code{trim_setsize} object.
}
\description{
Given a mass spectrometry design matrix and a gene pathways
  list (each within an \code{Omics.*}-class object), extract the genes /
  proteins / lipids / metabolomes contained in each gene pathway set which
  are expressed in the MS design matrix.
}
\details{
This function takes in a data frame with named columns and a pathway
  list, all through one of the \code{Omics.*} classes. This function will
  then iterate over the list of pathways, extract columns from the MS design
  matrix which match the genes listed in that pathway, and remove any
  pathways with fewer than \code{trim} expressed genes. The genes not
  expressed in the MS design matrix are removed from the pathway list.

  Note that some genes will be included in more than one pathway, so these
  pathways are not mutually exclusive. Further note that there may be many
  genes in the MS design matrix that are not included in the pathway sets, so
  these will not be extracted to the list. It is then vitally important to
  use either a very broad and generic pathway set list or a pathway set list
  that is appropriate for the mass spectrometry data supplied.
}
\examples{
\dontrun{
  ###  Load the Example Data  ###
  data("colonSurv_df")
  data("colonGenesets_ls")

  ###  Create an OmicsSurv Object  ###
  colon_OmicsSurv <- create_OmicsSurv(assayData_df = colonSurv_df[, -(1:2)],
                                      pathwaySet_ls = colonGenesets_ls,
                                      eventTime_vec = colonSurv_df$OS_time,
                                      eventObserved_vec = as.logical(colonSurv_df$OS_event))

  ###  Extract Expressed Genes  ###
  expressedOmes(colon_OmicsSurv)
}

}
