% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetExpressed-omes.R
\docType{methods}
\name{expressedOmes}
\alias{expressedOmes}
\alias{expressedOmes,OmicsPathway-method}
\title{Extract expressed -omes matching a gene set from a mass spectrometry matrix}
\usage{
expressedOmes(object, trim = 3, ...)

\S4method{expressedOmes}{OmicsPathway}(object, trim = 3, ...)
}
\arguments{
\item{object}{An object of class \code{OmicsPathway}, \code{OmicsSurv},
\code{OmicsReg}, or \code{OmicsCateg}}

\item{trim}{The minimum cutoff of expressed -ome measures before a pathway
is excluded. Defaults to 3.}

\item{...}{Dots for additional internal arguments (as necessary)}
}
\value{
A list of pathways with -ome measures expressed in the MS design
  matrix. Each element of the list will be named by its pathway, and the
  elements will be subset matrices of the original MS design matrix. See
  "details" for more information
}
\description{
Given a mass spectrometry design matrix and a gene pathways
  list (each within an \code{Omics.*}-class object), extract the genes /
  proteins / lipids / metabolomes contained in each gene pathway set which
  are expressed in the MS design matrix.
}
\details{
This function takes in a data frame with named columns and a pathway
  list, all through one of the \code{Omics.*} classes. This function will
  then iterate over the list of pathways, extract columns from the MS design
  matrix which match the genes listed in that pathway, and remove any
  pathways with fewer than \code{trim} expressed genes. These matrices are
  returned as a named list. Note that some genes will be included in more
  than one pathway, so these pathways are not mutually exclusive. Further
  note that there may be many genes in the MS design matrix that are not
  included in the pathway sets, so these will not be extracted to the list.
  It is then vitally important to use either a very broad and generic pathway
  set list or a pathway set list that is appropriate for the mass
  spectrometry data supplied.
}
