% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_matrixRoot.R
\name{matrixRoot}
\alias{matrixRoot}
\title{Positive Root of a Symmetric Matrix}
\usage{
matrixRoot(x, root = 2)
}
\arguments{
\item{x}{A symmetric (necessarily square) matrix}

\item{root}{A positive number}
}
\value{
A matrix, that when multiplied by itself \code{root} times, yields
  \code{x}.
}
\description{
Calculate the matrix root of a symmetric matrix via the Spectral
  Decomposition.
}
\details{
This function decomposes \code{x} into $V x D x V^T$ via the
  \code{eigen()} function, sets any numerically negative eigenvalues to 0,
  calculates the root of these eigenvalues as D^*, then returns the matrix
  $V x D^* x V^T$.
}
\examples{
  X <- matrix(rnorm(25), ncol = 5);   xTx <- t(X) \%*\% X
  matrixRoot(xTx)
  matrixRoot(xTx, root = 3)
}
