% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown_matrixNorm.R
\name{normalize}
\alias{normalize}
\title{Normalize a Matrix}
\usage{
normalize(B, d)
}
\arguments{
\item{B}{A matrix?}

\item{d}{The number of columns of \code{B} to normalize}
}
\value{
A modified version of the \code{B} matrix
}
\description{
A function that norms a matrix, but I don't understand any of
  it.
}
\details{
I met with James and Steven on 26 September and neither of them
  understood the sign reversal in the last line of the internal \code{for()}
  loop. Before this function can be extracted and documented, I need to
  understand why it's doing what it's doing. Based on how it's called in the
  \code{aespca()} function, it has something to do with adjusting the aes-pca
  eigenvectors returned by the \code{lars.lsa()} function.
}
\examples{
  NULL
}
