% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_model_LS.R
\name{olsTrain_fun}
\alias{olsTrain_fun}
\title{Gene-specific Regularized Ordinary Least Squares fit statistics for supervised PCA}
\usage{
olsTrain_fun(x, y, s0.perc = NULL)
}
\arguments{
\item{x}{An \eqn{p \times n} predictor matrix.}

\item{y}{A response vector.}

\item{s0.perc}{Percentile of the standard error of the slope estimate to be
used for regularization. The Default value of \code{NULL} will use the
median of this distribution.}
}
\value{
A list of OLS model statistics:
\itemize{
  \item{\code{tt} : }{The Student's \eqn{t} test statistic the slopes
     (\eqn{\beta}).}
  \item{\code{numer} : }{The estimate of \eqn{\beta}.}
  \item{\code{sd} : }{The standard error of the estimates for \eqn{\beta}
     (the standard error divided by the square root of Sxx).}
  \item{\code{fudge} : }{A regularization parameter. See Details for
     description.}
}
}
\description{
Model statistics for Ordinary Least Squares (OLS) regression by
  gene.
}
\details{
This function calculates the Sxx, Syy, and Sxy sums from the gene-
  specific OLS models, then calculates estimates of the regression slopes for
  each gene and their corresponding regularized test statistics,
  \deqn{t = \hat{\beta} / (sd + e),}
  where \eqn{e} is a regularization parameter.

  If \code{s0.perc} is \code{NULL}, then \eqn{e} is median of the \code{sd}
  values. Otherwise, \eqn{e} is set equal to \code{quantile(sd, s0.perc)}.
}
\examples{
  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use SuperPCA_pVals() instead
  
  
  p <- 500
  n <- 50

  x_mat <- matrix(rnorm(n * p), nrow = p, ncol = n)
  time_int <- rpois(n, lambda = 365 * 2)
  
  olsTrain_fun(
    x = x_mat,
    y = time_int
  )
  
  
}
\keyword{internal}
