% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClass_OmicsSurv.R
\docType{class}
\name{OmicsSurv-class}
\alias{OmicsSurv-class}
\title{An S4 class for survival responses within an "OmicsPathway" object}
\description{
This creates the "OmicsSurv" class which extends the
  "OmicsPathway" master class.
}
\section{Slots}{

\describe{
\item{\code{massSpec}}{An $N x p$ data frame with named columns}

\item{\code{pathwaySet}}{A list of known gene pathways with one or two elements:
\itemize{
  \item{pathways : }{A named list of character vectors. Each vector contains
    the names of the individual genes within that pathway as a vector of
    character strings. The names contained in these vectors must have non-
    empty overlap with the \emph{column names} of the \code{massSpec} data
    frame. The names of the pathways (the list elements themselves) should
    be the a shorthand representation of the full pathway name.}
  \item{TERMS: }{ A character vector the same length as the
    \code{pathways} list with the proper names of the pathways.}
  \item{setsize : }{A named integer vector the same length as the
    \code{pathways} list with the number of genes in each pathway. This list
    item is calculated during the creation step of a \code{create_Omics*()}
    function call.}
}}

\item{\code{eventTime}}{A numeric vector with $N$ observations corresponding to the
last observed time of follow up}

\item{\code{eventObserved}}{A logical vector with $N$ observations indicating right
censoring. The values will be FALSE if the observation was censored (i.e.,
we did not observe an event).}
}}

\seealso{
\code{"\link[=OmicsPathway-class]{OmicsPathway}"},
  \code{\link{create_OmicsSurv}}
}
