% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_pathway_pValues.R
\name{pathway_pValues}
\alias{pathway_pValues}
\title{Calculate the p-Values from a Mixture if Weibull Extreme Value Distributions}
\usage{
pathway_pValues(optimParams_vec, max_tScores_vec, genelist_ls,
  FDRadjust = TRUE, multTestProc = "BH")
}
\arguments{
\item{optimParams_vec}{A named vector of the estimated values for the
parameters which minimize the likelihood as returned by the function
\code{\link{weibullMix_optimParams}}.}

\item{max_tScores_vec}{A vector of the maximum absolute t-scores for each
pathway when under the alternative model (the response vector as is).}

\item{genelist_ls}{A list of three elements:
\itemize{
  \item{pathways : }{A list of character vectors such that each vector
     contains the ID numbers (as a character) of the individual genes
     within that pathway as a vector of character strings.}
  \item{TERMS : }{A character vector containing the names of the gene
     pathways.}
  \item{setsize : }{A named integer vector containing the number of genes
     in each gene pathway.}
}}

\item{FDRadjust}{Should the p-values be adjusted for multiple comparisons?
Defaults to \code{TRUE}.}

\item{multTestProc}{If the p-values should be adjusted, which procedure will
you use? Options are passed to the \code{\link[multtest]{mt.rawp2adjp}}
function. Defaults to "BH".}
}
\value{
A data frame
}
\description{
Calculate the pathway-specific p-values and their associated
   False Discovery Rates (FDR).
}
\details{
This function takes in the optimal parameters returned by the
   \code{\link{weibullMix_optimParams}} function, the maximum t-Scores for
   each gene pathway, and the list of gene pathway information. This function
   will calculate the p-value for each t-Score given the Gumbel Extreme Value
   mixture distribution parametrized by the values of \code{optimParams}. If
   requested, this function will also calculate the FDR associated with all
   pathway p-values via the Benjamini & Hochberg (1995) step-up FDR-
   controlling procedure as implemented in the \code{multtest} package.
}
\examples{
   NULL
}
