% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_pathway_tScores.R
\name{pathway_tScores}
\alias{pathway_tScores}
\title{Calculate Pathway-Specific Student's t Scores}
\usage{
pathway_tScores(pathway_vec, geneArray_df, response_mat,
  responseType = "survival", n.threshold = 20, numPCs = 1,
  min.features = 5)
}
\arguments{
\item{pathway_vec}{A character vector of the measured -omes in the chosen
gene pathway. These should match a subset of the rownames of the gene
array.}

\item{geneArray_df}{A "tall" pathway data frame ($p * n$). Each subject or
tissue sample is a column, and the rows are the -ome measurements for that
sample.}

\item{response_mat}{A response matrix corresponding to \code{responseType}.
For \code{"regression"} and \code{"classification"}, this will be an
$n * 1$ matrix of response values. For \code{"survival"}, this will be an
$n * 2$ matrix with event times in the first column and censoring
indicator in the second.}

\item{responseType}{A character string. Options are \code{"survival"},
\code{"regression"}, and \code{"classification"}.}

\item{n.threshold}{The number of bins into which to split the feature scores
in the \code{fit} object returned internally by the \code{superpc.train}
function.}

\item{numPCs}{The number of PCs to extract from the significant pathway.}

\item{min.features}{What is the smallest number of genes allowed in each
pathway? This argument must be kept constant across all calls to this
function which use the same pathway list. Defaults to 5.}
}
\value{
A matrix with \code{numPCs} rows and \code{n.threshold} columns:
   model t-statisics for each PC included (rows) at each threshold level
   (columns).
}
\description{
Extract principal components from the gene pathway, and return
   the test statistics associated with the first \code{numPCs} principal
   components at a set of threshold values.
}
\details{
This is a wrapper function to call \code{\link{superpc.train}}
   and \code{\link{superpc.st}}. This wrapper is designed to facilitate apply
   calls (in parallel or serially) of these two functions over a list of gene
   pathways. When \code{numPCs} is 1, we recommend using a simplify-style
   apply variant, such as \code{\link[base]{sapply}} or
   \code{\link[parallel]{parSapply}}, then transposing the resulting matrix.
}
\examples{
   NULL
}
\seealso{
\code{\link{superpc.train}}; \code{\link{superpc.st}}
}
