% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesPC_permtest_GLM.R
\docType{methods}
\name{PermTestCateg}
\alias{PermTestCateg}
\alias{PermTestCateg,OmicsCateg-method}
\title{AES-PCA permutation test of categorical response for pathway PCs}
\usage{
PermTestCateg(OmicsCateg, pathwayPCs_ls, numReps = 1000,
  parallel = FALSE, numCores = NULL, ...)

\S4method{PermTestCateg}{OmicsCateg}(OmicsCateg, pathwayPCs_ls,
  numReps = 1000, parallel = FALSE, numCores = NULL, ...)
}
\arguments{
\item{OmicsCateg}{A data object of class \code{OmicsCateg}, created by the
\code{\link{CreateOmicsCateg}} function.}

\item{pathwayPCs_ls}{A list of pathway PC matrices returned by the
\code{\link{ExtractAESPCs}} function.}

\item{numReps}{How many permuted models to fit? Defaults to 1000.}

\item{parallel}{Should the computation be completed in parallel? Defaults to
\code{FALSE}.}

\item{numCores}{If \code{parallel = TRUE}, how many cores should be used for
computation? Defaults to \code{NULL}.}

\item{...}{Dots for additional internal arguments (currently unused).}
}
\value{
A named vector of pathway permutation \eqn{p}-values.
}
\description{
Given an \code{OmicsCateg} object and a list of pathway PCs from
  the \code{\link{ExtractAESPCs}} function, test if each pathway with
  features recorded in the bio-assay design matrix is significantly related
  to the categorical response.
}
\details{
This function takes in a list of the first principal components
  from each pathway and an object of class \code{OmicsCateg}. This function
  will then calculate the AIC of a multivariate generalized linear model (via
  the \code{\link[stats]{glm}} function with a \code{\link[stats]{binomial}}
  error family) with the original observations as response and the pathway
  principal components as the predictor matrix.

  Then, this function will create \code{numReps} permutations of the
  categorical response, fit models to each of these permuted responses
  (holding the path predictor matrix fixed), and calculate the AIC of each
  model. This function will return a named vector of permutation
  \eqn{p}-values, where the value for each pathway is the proportion of
  models for which the AIC of the permuted response model is less than the
  AIC of the original model.

  In future versions, this function will also be able to calculate permuted
  \eqn{p}-values for multinomial logistic regression and proportional odds
  logistic regression models, for n-ary and ordered categorical responses,
  respectively.
}
\examples{
  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use AESPCA_pVals() instead

}
\seealso{
\code{\link{CreateOmicsCateg}}; \code{\link{ExtractAESPCs}};
  \code{\link[stats]{glm}}; \code{\link[stats]{binomial}};
  \code{\link{SampleCateg}}
}
\keyword{internal}
