% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_model_tStats.R
\name{superpc.st}
\alias{superpc.st}
\title{Extract and Test Supervised PCs}
\usage{
superpc.st(fit, data, n.threshold = 20, n.PCs = 1, min.features = 5,
  epsilon = 1e-06)
}
\arguments{
\item{fit}{An object of class \code{superpc} returned by the function
\code{superpc.train}}

\item{data}{A list of testing data:
\itemize{
  \item{x : }{A "tall" pathway data frame ($p_{path} * n$).}
  \item{y : }{A response vector corresponding to \code{type}}
  \item{censoring.status : }{If \code{type = "survival"}, the censoring
    indicator. Otherwise, \code{NULL}.}
  \item{featurenames : }{A character vector of the measured -omes in
    \code{x}.}
 }}

\item{n.threshold}{The number of bins into which to split the feature scores
returned in the \code{fit} object.}

\item{n.PCs}{The number of PCs to extract from the significant pathway}

\item{min.features}{What is the smallest number of genes allowed in each
pathway? This argument must be kept constant across all calls to this
function which use the same pathway list. Defaults to 5}

\item{epsilon}{I'm not sure why this is important. It's called when comparing
the absolute score values to each value of the threshold vector. Defaults
to 10 ^ -6.}
}
\value{
A list containing:
\itemize{
  \item{thresholds : }{A labelled vector of quantile values of the score
    vector in the \code{fit} object.}
  \item{n.threshold : }{The number of splits to make in the score vector.}
  \item{scor : }{A matrix of model fit statistics. Each column is the
    threshold level of predictors allowed into the model, and each row is a
    PC included. Which genes are included in the matrix before PC extraction
    is governed by comparing their model score to the quantile value of the
    scores at each threshold value.}
  \item{tscor : }{A matrix of model t-statisics for each PC included (rows)
    at each threshold level (columns).}
  \item{type : }{Which model was called? Options are survival, regression,
    or binary.}
}
}
\description{
Identify significant features, extract PCs from those specific
   features to construct a data matrix, predict the response with this data
   matrix, and record the model fit statistic of this prediction.
}
\details{
See \url{https://web.stanford.edu/~hastie/Papers/spca_JASA.pdf}
  An issue, the number of thresholds at which to test (\code{n.threshold}),
  can be larger than the number of features to bin. This is why so many of
  the t-statistics are constant - because the model isn't changing.
}
\examples{
  NULL
}
