% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_train.R
\name{superpc.train}
\alias{superpc.train}
\title{Train Supervised PC Model}
\usage{
superpc.train(data, type = c("survival", "regression", "binary"),
  s0.perc = NULL)
}
\arguments{
\item{data}{A list of training data:
\itemize{
  \item{x : }{A "tall" pathway data frame ($p_{path} * n$).}
  \item{y : }{A response vector corresponding to \code{type}}
  \item{censoring.status : }{If \code{type = "survival"}, the censoring
    indicator. Otherwise, \code{NULL}.}
  \item{featurenames : }{A character vector of the measured -omes in
    \code{x}.}
}}

\item{type}{What model relates \code{y} and \code{x}? Options are
\code{"survival"}, \code{"regression"}, or \code{"binary"} (for logistic
regression).}

\item{s0.perc}{A stabilization parameter on the interval [0,1]. This is an
internal argument to each of the called functions. The default NULL value
will ensure an appropriate value is determined internally.}
}
\value{
A list containing:
\itemize{
  \item{feature.scores : }{The scaled p-dimensional score vector: each value
    has been divided by its respective standard deviation plus epsilon
    (governed by \code{s0.perc}). \code{NA} values returned by the logistic
    model are replaced with 0.}
  \item{type : }{The argument for \code{type}.}
  \item{s0.perc : }{The user-supplied value of \code{s0.perc}, or the
    internally-calculated default value from the chosen model.}
  \item{call : }{The output of \code{match.call()}.}
}
}
\description{
Computes feature scores for supervised pc analysis
}
\details{
This function is a switch call to \code{\link{coxTrain_fun}},
   \code{cor.func}, or \code{logistic.func}, respectively.
}
\examples{
   NULL
}
