% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClass_validOmics.R
\name{valid_OmicsSurv}
\alias{valid_OmicsSurv}
\alias{valid_OmicsReg}
\alias{valid_OmicsCateg}
\title{Validity Checking for -Omics.* Classes}
\usage{
valid_OmicsSurv(object)

valid_OmicsReg(object)

valid_OmicsCateg(object)
}
\arguments{
\item{object}{An object of classes "OmicsSurv", "OmicsReg", or "OmicsCateg".}
}
\value{
TRUE if the object is a valid object, else an error message with the
  rule broken.
}
\description{
These functions check the validity of the "OmicsSurv", "OmicsReg", and
  "OmicsCateg" classes.
}
\details{
At the moment, we have currently written checks to make sure the dimensions
  of the mass spectrometry or bio-assay data frame and response vectors
  match.
}
\section{OmicsSurv}{

Valid OmicsSurv objects will have two response vectors: a vector of the most
  recently recorded follow-up times and a logical vector if that time marks
  an event (TRUE = observed event; FALSE = right-censored observation).
}

\section{OmicsReg and OmicsCateg}{

Valid OmicsReg and OmicsCateg objects with have one response vector of
  continuous or categorial (as a factor) observations, respectively.
}

