% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC_optimWeibull_pValues.R
\name{weibullMix_pValues}
\alias{weibullMix_pValues}
\title{Calculate the p-Values from an Optimal Mixture of Weibull Distributions}
\usage{
weibullMix_pValues(tScore_vec, pathwaySize_vec, optimParams_vec)
}
\arguments{
\item{tScore_vec}{A vector of the maximum absolute t-scores for each pathway
when under the alternative model}

\item{pathwaySize_vec}{A vector of the number of genes in each pathway}

\item{optimParams_vec}{The NAMED vector of optimal mixture distribution
parameters returned by the \code{\link{weibullMix_optimParams}} function.}
}
\value{
A named vector of the estimated raw p-values for each gene pathway
}
\description{
Calculate the p-values of test statistics from a mixture of two
  Weibull Extreme Value distributions.
}
\details{
The likelihood function is equation (4) Chen et al (2008): a mixture
 of two Gumbel Extreme Value pdfs, with mixing proportion p. The values mu_i
 and s_i, i = 1, 2, within the code of the function are placeholders for the
 mean and standard deviation, respectively.

 @seealso \code{\link[stats]{optim}}
}
\examples{
   NULL
}
